+++ src/components/CardList.js
@@ -1,7 +1,11 @@
 import React from 'react';
 import Card from './Card';
 
+
 const CardList = ({ robots }) => {
+  // if (true) {
+  //   throw new Error('NOOOO!')
+  // }
     const cardArray = robots.map((user, i) => {
          return (
          <Card key={i}
