+++ src/components/ErrorBoundry.js
@@ -1,0 +1,24 @@
+import React, { Component } from 'react';
+
+class ErrorBoundry extends Component {
+    constructor(props){
+        super(props);
+        this.state = {
+            hasError: false
+        }
+    }
+
+    componentDidCatch(error,info){
+        this.setState({ hasError: true})
+    }
+
+    render() {
+      if (this.state.hasError) {
+          return <h1> Ops, esto no es nada bueno!</h1>
+      }
+      return this.props.children
+    }
+
+}
+
+export default ErrorBoundry;
