+++ src/conteiners/App.js
@@ -1,37 +1,50 @@
 import React, { Component } from 'react';
 import CardList from '../components/CardList';
-import { robots } from '../components/robots';
+
 import SearchBox from '../components/SearchBox';
 import Scroll from '../components/Scroll'
+import ErrorBoundry from '../components/ErrorBoundry';
 
 class App extends Component {
     constructor() {
         super()             //llama al constructor
         this.state = {
-            robots: robots,   //ahora robots es parte del objeto, de la propiedad state
+            robots: [],   //se crea, y luego que los datos dl servidor se monten lo garga
             searchfield: '',
         }
     }
+
+    componentDidMount() {
+        fetch('https://jsonplaceholder.typicode.com/users')
+        .then(response=> response.json())   //json(), transforma lo que recibe en archivo json
+        .then(users => this.setState({ robots:users })); 
+    }
+
     onSearchChange = (event) => {
         this.setState({ searchfield: event.target.value })
     }
     
     render () {
-        const { robots, searchfield } = this.state;   
-        const filteredRobots= robots.filter(robot=> {
-            return robot.name.toLowerCase().includes(searchfield.toLowerCase())
+        const filteredRobots= this.state.robots.filter(robot=> {
+            return robot.name.toLowerCase().includes(this.state.searchfield.toLowerCase())
         })
-       return !robots.length 
-        ?  <h1> Cargando ... </h1> 
-        : (<div className='tc'>
-                     <h1 style= {{color: 'black'}}>RoboFriends</h1>
-                     <SearchBox searchChange={this.onSearchChange}/>
-                     <Scroll>
-                        <CardList robots={filteredRobots}/>   
-                     </Scroll>
-          </div>);
-      }   
- }
+        if(this.state.robots.length === 0) {
+            return <h1> Loaring ... </h1>
+        } else {
+            return (
+                <div className='tc'>
+                    <h1 style= {{color: 'black'}}>RoboFriends</h1>
+                    <SearchBox searchChange={this.onSearchChange}/>
+                    <Scroll>
+                        <ErrorBoundry>
+                          <CardList robots={filteredRobots}/>   
+                        </ErrorBoundry>
+                    </Scroll>
+                </div>
+          );
+      }
+   }
+}
 
 export default App;
